Rewrite play card
