+++ love_letter/usecase/play_card.py
@@ -1,9 +1,7 @@
 import traceback
-from dataclasses import dataclass
 from typing import Union
 
 from love_letter.models import Game, GuessCard, ToSomeoneCard
-from love_letter.models.event import DomainEvent
 from love_letter.usecase.common import Presenter, game_repository
 
 
@@ -14,19 +12,6 @@ class PlayCardInput:
     card_action: Union[GuessCard, ToSomeoneCard, None]
 
 
-@dataclass
-class CardPlayedEvent(DomainEvent):
-    game: Game
-
-
-class PlayCardPresenter(Presenter):
-    def as_view_model(self) -> Game:
-        for event in self.events:
-            if isinstance(event, CardPlayedEvent):
-                return event.game
-        raise BaseException("Game is unavailable.")
-
-
 class PlayCard:
     def execute(self, input: PlayCardInput, presenter: Presenter):
         try:
@@ -34,9 +19,9 @@ class PlayCard:
             if game is None:
                 raise BaseException(f"Game {input.game_id} is unavailable.")
 
-            game.play(input.player_id, input.card_name, input.card_action)
+            event = game.play(input.player_id, input.card_name, input.card_action)
             game_repository.save_or_update(game)
-            presenter.present([CardPlayedEvent(game)])
+            presenter.present(event)
         except BaseException as exception:
             traceback.print_exception(exception)
 
@@ -54,7 +39,3 @@ class PlayCard:
         input.card_name = card_name
         input.card_action = card_action
         return input
-
-    @classmethod
-    def presenter(cls) -> "PlayCardPresenter":
-        return PlayCardPresenter()
